Remove api and client module, and duplicated method in args.py from main package
