+++ elasticdl/python/elasticdl/local_executor.py
@@ -17,7 +17,6 @@ from collections import namedtuple
 
 import tensorflow as tf
 
-from elasticdl.python.common.args import parse_envs
 from elasticdl.python.common.constants import MetricsDictKey
 from elasticdl.python.common.evaluation_utils import EvaluationMetrics
 from elasticdl.python.common.log_utils import default_logger as logger
@@ -29,6 +28,7 @@ from elasticdl.python.data.reader.csv_reader import CSVDataReader
 from elasticdl.python.data.reader.data_reader_factory import create_data_reader
 from elasticdl.python.data.reader.odps_reader import ODPSDataReader
 from elasticdl.python.data.reader.recordio_reader import RecordIODataReader
+from elasticdl_client.common.args import parse_envs
 
 _MockedTask = namedtuple("Task", ["shard_name", "start", "end"])
