+++ elasticdl/python/master/k8s_instance_manager.py
@@ -17,8 +17,9 @@ import threading
 from collections import Counter
 
 from elasticdl.python.common import k8s_client as k8s
-from elasticdl.python.common.constants import BashCommandTemplate, PodStatus
+from elasticdl.python.common.constants import PodStatus
 from elasticdl.python.common.log_utils import default_logger as logger
+from elasticdl_client.common.constants import BashCommandTemplate
 
 _SERVICE_ADDR_SEP = ","
