+++ elasticdl/python/tests/args_test.py
@@ -11,46 +11,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import argparse
 import unittest
 
 from elasticdl.python.common.args import (
-    add_bool_param,
-    build_arguments_from_parsed_result,
     parse_ps_args,
     wrap_go_args_with_string,
-    wrap_python_args_with_string,
 )
 
 
 class ArgsTest(unittest.TestCase):
-    def setUp(self):
-        self._parser = argparse.ArgumentParser()
-        self._parser.add_argument("--foo", default=3, type=int)
-        add_bool_param(self._parser, "--bar", False, "")
-
-    def test_build_arguments_from_parsed_result(self):
-        args = ["--foo", "4", "--bar"]
-        results = self._parser.parse_args(args=args)
-        original_arguments = build_arguments_from_parsed_result(results)
-        value = "\t".join(sorted(original_arguments))
-        target = "\t".join(sorted(["--foo", "4", "--bar", "True"]))
-        self.assertEqual(value, target)
-
-        original_arguments = build_arguments_from_parsed_result(
-            results, filter_args=["foo"]
-        )
-        value = "\t".join(sorted(original_arguments))
-        target = "\t".join(sorted(["--bar", "True"]))
-        self.assertEqual(value, target)
-
-        args = ["--foo", "4"]
-        results = self._parser.parse_args(args=args)
-        original_arguments = build_arguments_from_parsed_result(results)
-        value = "\t".join(sorted(original_arguments))
-        target = "\t".join(sorted(["--bar", "False", "--foo", "4"]))
-        self.assertEqual(value, target)
-
     def test_parse_ps_args(self):
         minibatch_size = 16
         num_minibatches_per_task = 8
@@ -82,26 +51,6 @@ class ArgsTest(unittest.TestCase):
         )
         self.assertEqual(parsed_args.model_zoo, model_zoo)
         self.assertEqual(parsed_args.model_def, model_def)
-
-    def test_wrap_python_args_with_string(self):
-        args = [
-            "--ps_id",
-            str(0),
-            "--job_name",
-            "test_args",
-            "--checkpoint_dir",
-            "",
-        ]
-        args = wrap_python_args_with_string(args)
-        expected_args = [
-            "--ps_id",
-            "'0'",
-            "--job_name",
-            "'test_args'",
-            "--checkpoint_dir",
-            "''",
-        ]
-        self.assertListEqual(args, expected_args)
 
     def test_wrap_go_args_with_string(self):
         args = [
